Added summary for ASTRA ("Transforming signatures...").
FACT:
Asociaţia Transilvană pentru Literatura Română şi Cultura Poporului Român (ASTRA; English: The Transylvanian Association for Romanian Literature and the Culture of the Romanian People) is a cultural association founded in 1861 in Sibiu. It had an important role in the cultural life and the movement of national awakening for the Romanians in Transylvania.