New FilterType/FilterAction approach build in Core Compiletime: No primitive Filtertypes anymore. A filtertype exists only when the corresponding FilterType object is present in the repository.
FilterTypes now consists of 4 different FilterActions: accept-call-action, reject-call-action, accept-return-action, reject-return-action. These are references to FilterAction objects in the Repository

The platform-dependent compiletime is responsible for creating the FilterType and FilterAction objects in the repository for every FilterType and FilterAction.

Although there are no primitive filtertypes/filteractions anymore, the core still depends on some build-in filteractions to be present, especially the DispatchAction and the ContinueAction