Added NUnit. The tests can now be run in either VSTS or NUnit. Just open the correct solution file. The StarLight.sln contains only source code. The StarLight.VSTS.sln contains source code and VSTS tests. The StarLight.NUnit.sln contains source code and NUnit tests.
When writing the unit tests, make sure it works in both systems. Supply two project files (a namespace.Tests.NUnit and a namespace.Tests.VSTS). In the NUnit project, set a compiler compilation symbol NUNIT and reference the NUnit.Framework.dll

Do not use the VSTS auto generated test option. This will generate a stub for accessing the test target and this will not work with NUnit. 

Also in this revision; cleanup of old binaries no longer used. small fix in the EntitiesAccessor.