removed classPath and mainclass from the setup and settings classPath
created new settings container, StarLightSettings.cs based on the ApplicationSettingsBase class.
Created registry settings provider. Due to load problems, this class is located in the entities assembly. Could not find a better solution for this.

The java debug levels now use strings instead of numbers in the cpsproj files to make it consistent with the weavedebuglevels.

Settings for the java part are now stored in a setting list in the configContainer for better extensibility. 

Added option to disable the FILTH output. Default is enabled.