Added the following custom attributes:  Use the SkipWeaving attribute to exclude elements from the weaving process. Like complete assemblies, classes, methods.  The ProcessProperties allows the developer to enable the processing and thus weaving on properties. default action is disabled and it can be enabled on assembly and class level.
The SkipWeaving attribute is injected into an assembly StarLight has weaved on. This makes sure we are not weaving on an assembly multiple times. 

Note; properties are, by default now, not processed. So no code is generated and thus added inside properties. This can improve runtime performance since the JITter can now inline properties again.