real(1) regular expression patterns can now be used in FIRE2's regex util.
(1) only a limited feature set is accepted. It works on words (sequence of characters) rather than single characters. 
Quantifiers (?, * or +) are only allowed for subexpressions (e.g. (foo)*). 
The dot (match one character) can only be combined with the * or + quantifier (.* or .+; meaning zero or many words, or 1 or many words). 
Alternations are created using the pipe. 
The following regex are equal: foo|bar|quux ; (foo)|(bar)|quux ; (foo|bar)|(quux)
Sequences are simply created by concatination, but be aware of the word principle: "readwrite" is read as a single word. Use: "(read)(write)" for two words.
Negation is possible with the "Zero-width negative lookahead" constructions: (?!foo)
To match anything except foo and bar: (?!foo|bar) or (?!(foo|bar)) or (?!(foo)|(bar))
Negating of sequences is broken. (?!(foo)(bar)) does NOT work (yet).
The reason for this is that the inverse the sequence is not yet created (because it's complex, specially with nested negations).
"foo|(?!bar|(?!quux))" equals "foo|(?!bar)|quux", which is of course identical to "(?!bar)", but this kind of optimization isn't performed.